Move catalog handler out of download protocol
